Update WMIMonitor.ps1
Reworked the New-EventSubscriberMonitor function to get more consistent results. Basically got rid of nested hashtables, and added more explicit casting for a lot of the arguments. Tested on Win7 and Win8 systems. Works more consistently, although I did see some issues with a couple of Win7 builds - but not as many as the original script. Worked on all Win8 systems I tested on.